pyproject.toml
[tool.poetry]
name = "prisjakt-scraper"
version = "0.1.0"
description = ""
authors = ["Harald Vinje <haraldvinje@gmail.com>"]
packages = [{ include = "prisjakt_scraper" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
lxml = "^4.9.2"
unidecode = "^1.3.6"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^2.0.9"
pandas = "^2.0.0"
alembic = "^1.10.2"
python-dotenv = "^1.0.0"
boto3 = "^1.26.104"
pytest = "^7.2.2"
pytz = "^2023.3"
pytest-mock = "^3.10.0"
selenium = "^4.12.0"
webdriver-manager = "^4.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.3"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
ignore = ['E231', 'E241']
per-file-ignores = ['__init__.py:F401']
max-line-length = 120
count = true

[tool.pylint.'linters']
disable = [
    "missing-docstring, logging-format-interpolation",
    "logging-fstring-interpolation",
    "too-many-instance-attributes",
    "too-few-public-methods",
    "not-callable"
]

[tool.pylint.basic]
good-names = ["ad", "i"]
